# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inventory adjustments dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals/global:
    put:
      operationId: setGlobalGoal
      tags: [goals]
      summary: Set global min/max goal band
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoalBandInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GoalBand" }
  /goals/stock/{stockId}:
    put:
      operationId: setStockGoal
      tags: [goals]
      parameters:
        - name: stockId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoalBandInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockGoal" }
    delete:
      operationId: deleteStockGoal
      tags: [goals]
      parameters:
        - name: stockId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
  /snapshots/{weekEnding}:
    get:
      operationId: getWeeklySnapshot
      tags: [snapshots]
      summary: Get a single snapshot with roll details
      parameters:
        - name: weekEnding
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Snapshot detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeeklySnapshot" }
...
  /monthly-snapshots/{monthKey}:
    get:
      operationId: getMonthlySnapshot
      tags: [snapshots]
      summary: Get a single monthly snapshot with roll details
      parameters:
        - name: monthKey
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Monthly snapshot detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonthlySnapshot" }
...
components:
...
  schemas:
...
    GoalBand:
      type: object
      required: [min, max]
      properties:
        min: { type: number, nullable: true }
        max: { type: number, nullable: true }
        serviceLevel:
          type: number
          nullable: true
          description: Shared Demand Planning service-level default (0–1). Null = app default 0.95.
        monthsBack:
          type: integer
          nullable: true
          description: Shared Demand Planning history-window default in months. Null = app default 6.
        demandCv:
          type: number
          nullable: true
          description: Shared Demand Planning demand-CV override default. Null = no override.
        leadTimeCv:
          type: number
          nullable: true
          description: Shared Demand Planning lead-time-CV override default. Null = no override.
    GoalBandInput:
      type: object
      properties:
        min: { type: number, nullable: true }
        max: { type: number, nullable: true }
        demandCv: { type: number, nullable: true }
        leadTimeCv: { type: number, nullable: true }
        serviceLevel: { type: number, nullable: true }
        monthsBack: { type: integer, nullable: true }
        seasonalityWeights:
          type: array
          nullable: true
          minItems: 3
          maxItems: 3
          items: { type: number }
          description: Quarterly seasonality weights for months 1, 2, 3 of each quarter. Should sum to ~1.0.
        leadTimeDays:
          type: number
          nullable: true
          description: Manual override for the average supplier lead time in days. Null = use observed PO history.
        typicalRollFootage:
          type: number
          nullable: true
          description: Manual override for the typical incoming roll size (footage). Null = use observed roll history.
    StockGoal:
      type: object
      required: [stockId, min, max]
      properties:
        stockId: { type: string }
        min: { type: number, nullable: true }
        max: { type: number, nullable: true }
        demandCv: { type: number, nullable: true }
        leadTimeCv: { type: number, nullable: true }
        seasonalityWeights:
          type: array
          nullable: true
          minItems: 3
          maxItems: 3
          items: { type: number }
        leadTimeDays: { type: number, nullable: true }
        typicalRollFootage: { type: number, nullable: true }
...
    SnapshotSummary:
      type: object
      required: [id, weekStart, weekEnding, capturedAt, onHandValue, rollCount, added, removed, netAdjustment, adjustmentPct]
      properties:
        id: { type: string }
        weekStart: { type: string, format: date }
        weekEnding: { type: string, format: date }
        capturedAt: { type: string, format: date-time }
        onHandValue: { type: number }
        rollCount: { type: integer }
        added: { type: number }
        removed: { type: number }
        netAdjustment: { type: number }
        adjustmentPct: { type: number }
    SnapshotRoll:
      type: object
      required: [direction, rollTag, stockId, amount, ccDate, ccString]
      properties:
        direction:
          type: string
          enum: [added, removed]
        rollTag: { type: string }
        stockId: { type: string }
        description: { type: string, nullable: true }
        amount: { type: number }
        ccDate: { type: string, format: date }
        ccString: { type: string }
    WeeklySnapshot:
      allOf:
        - $ref: "#/components/schemas/SnapshotSummary"
        - type: object
          required: [rolls]
          properties:
            rolls:
              type: array
              items: { $ref: "#/components/schemas/SnapshotRoll" }
    MonthlySnapshotSummary:
      type: object
      required: [id, monthKey, monthStart, monthEnd, capturedAt, onHandValue, rollCount, added, removed, netAdjustment, adjustmentPct]
      properties:
        id: { type: string }
        monthKey: { type: string }
        monthStart: { type: string, format: date }
        monthEnd: { type: string, format: date }
        capturedAt: { type: string, format: date-time }
        onHandValue: { type: number }
        rollCount: { type: integer }
        added: { type: number }
        removed: { type: number }
        netAdjustment: { type: number }
        adjustmentPct: { type: number }
    MonthlySnapshot:
      allOf:
        - $ref: "#/components/schemas/MonthlySnapshotSummary"
        - type: object
          required: [rolls]
          properties:
            rolls:
              type: array
              items: { $ref: "#/components/schemas/SnapshotRoll" }
...